fix circular dependency - move util function to renderer package
