+++ packages/ui/src/utils/helpers.tsx
@@ -1,11 +1,7 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import BigNumber from 'bignumber.js';
-import { chainCurrency, chainUnits } from '@ren/config/chains';
-import { planckToUnit, rmCommas } from '@w3ux/utils';
 import { LoadingPlaceholderWrapper } from '../styles/LoadingPlaceholderWrapper';
-import type { ChainID } from '@polkadot-live/types/chains';
 
 /**
  * @name createArrayWithLength
@@ -13,23 +9,6 @@ import type { ChainID } from '@polkadot-live/types/chains';
  */
 export const createArrayWithLength = (n: number): number[] =>
   [...Array(n + 1)].map((_, i) => i);
-
-/**
- * @name formatChainUnits
- * @summary Get readable chain units for rendering.
- */
-export const formatChainUnits = (units: string, chainId: ChainID) => {
-  // Include regex to remove trailing zeros after decimal point.
-  const formatted: string = planckToUnit(
-    new BigNumber(rmCommas(units)),
-    chainUnits(chainId)
-  )
-    .toFixed(2)
-    .replace(/(\.\d*?[1-9])0+|\.0*$/, '$1')
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-
-  return `${formatted} ${chainCurrency(chainId)}`;
-};
 
 /**
  * @name renderPlaceholders
@@ -53,7 +32,6 @@ export const renderPlaceholders = (
  * @name validateAccountName
  * @summary Validate an account name received from user input.
  */
-
 export const validateAccountName = (accountName: string): boolean => {
   // Regulare expression for allowed characters in the account name (including spaces).
   const regex = /^[a-zA-Z0-9._-\s]+$/;
